Proposal for Reward Standards
Thank you for your continued support and contributions over the past year since DVP was founded. With "Blockchain Security Evolution" global roadshows and hackathons, DVP has entered a new era of development! Lately, we received lots of feedback from white hat hackers in the DVP community with regards to the current Reward Standards 2.0, and after thorough investigation by the DVP team, we believe that the Reward Standards 2.0 are no longer suitable and requires immediate upgrade. In order to better guide our community, we have decided to release a proposal for Reward Standards 3.0